+++ src/ledger/persistence/SQLCurrencyRepository.hpp
@@ -16,6 +16,9 @@
 
 #pragma once
 
+// Standard include(s)
+#include <cstdint>
+
 // Qt include(s)
 #include <QSqlDatabase>
 #include <QString>
@@ -51,7 +54,7 @@ public:
      * @param[in] currency Currency to be created
      * @return ID assigned to the newly created currency
      */
-    int create(const Currency & currency) override;
+    int64_t create(const Currency & currency) override;
 
     /**
      * Retrieves a currency from the repository with the given ID.
@@ -60,7 +63,7 @@ public:
      * @return Currency for the given ID
      * @throw std::invalid_argument if no currency exists for the given ID
      */
-    Currency getCurrency(int id) override;
+    Currency getCurrency(int64_t id) override;
 
     /**
      * Returns the message of the last error to have occurred.
